requirements.txt
aiofiles==25.1.0
aiogram==3.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.22.0
annotated-types==0.7.0
attrs==25.4.0
beautifulsoup4==4.12.2
certifi==2025.11.12
charset-normalizer==3.4.4
frozenlist==1.8.0
idna==3.11
langdetect==1.0.9
lxml==4.9.3
magic-filter==1.0.12
multidict==6.7.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.0
requests==2.31.0
six==1.17.0
soupsieve==2.8.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
yarl==1.22.0
